chore: moved nav to comp folder structure
